backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# File Parsing
pdfplumber==0.10.3
python-docx==1.1.0
docx2txt==0.8

# NLP & AI
openai==1.3.5
spacy==3.7.2
transformers==4.35.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Report Generation
reportlab==4.0.7
weasyprint==60.1

# Optional: For better text processing
nltk==3.8.1
python-jose[cryptography]==3.3.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Security & Rate Limiting
slowapi==0.1.9

# Authentication & Database
bcrypt==4.1.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
httpx==0.25.2
google-auth==2.23.4

